+++ main/main-server/src/main/java/ru/practicum/ewm/main/server/entity/Event.java
@@ -86,9 +86,6 @@ public class Event {
     @Enumerated(EnumType.STRING)
     private EventStatus state;
 
-    @Column(name = "views")
-    private Long views;
-
     @Column(name = "confirmed_requests")
     private Long confirmedRequests;
 }
