+++ main/main-server/src/main/java/ru/practicum/ewm/main/server/service/EventService.java
@@ -1,19 +1,11 @@
 package ru.practicum.ewm.main.server.service;
 
-import ru.practicum.ewm.main.application.event.EventFullDto;
-import ru.practicum.ewm.main.application.event.EventShortDto;
-import ru.practicum.ewm.main.application.event.EventUpdateAdminRequest;
-import ru.practicum.ewm.main.application.event.EventUpdateUserRequest;
-import ru.practicum.ewm.main.application.event.NewEventDto;
-import ru.practicum.ewm.main.application.event.RequestStatusUpdateRequest;
-import ru.practicum.ewm.main.application.event.RequestStatusUpdateResult;
-import ru.practicum.ewm.main.application.request.ParticipationRequestDto;
+import ru.practicum.ewm.main.application.event.*;
 import ru.practicum.ewm.main.server.model.EventGetAllByAdminParameters;
 import ru.practicum.ewm.main.server.model.EventGetAllParameters;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.Collection;
-import java.util.List;
 
 public interface EventService {
 
@@ -32,8 +24,6 @@ public interface EventService {
     EventFullDto getByIdByPublic(long eventId, HttpServletRequest httpRequest);
 
     EventFullDto updateByAdmin(EventUpdateAdminRequest eventUpdateAdminRequest, long eventId);
-
-    List<ParticipationRequestDto> getRequests(Long eventId, Long userId);
 
     RequestStatusUpdateResult updateRequestStatus(RequestStatusUpdateRequest dto, Long eventId, Long userId);
 }
