+++ main/server-util/src/main/java/ru/practicum/ewm/main/util/exception/AlreadyUsedException.java
@@ -1,0 +1,7 @@
+package ru.practicum.ewm.main.util.exception;
+
+public class AlreadyUsedException extends RuntimeException {
+    public AlreadyUsedException(String message) {
+        super(message);
+    }
+}
